Module Lp_glpk.Milp

val solve : ?term_output:bool -> ?msg_lev:T.Msg.t option -> ?meth:T.Smcp.Meth.t option -> ?pricing:T.Smcp.Pt.t option -> ?r_test:T.Smcp.Rt.t option -> ?it_lim:int option -> ?tm_lim:int option -> ?br_tech:T.Iocp.Br.t option -> ?bt_tech:T.Iocp.Bt.t option -> ?pp_tech:T.Iocp.Pp.t option -> ?sr_heur:bool option -> ?fp_heur:bool option -> ?ps_heur:bool option -> ?ps_tm_lim:int option -> ?gmi_cuts:bool option -> ?mir_cuts:bool option -> ?cov_cuts:bool option -> ?clq_cuts:bool option -> ?tol_int:float option -> ?tol_obj:float option -> ?mip_gap:float option -> ?tm_lim_int:int option -> Lp.Problem.t -> (float * float Lp.PMap.t, string) Stdlib.result

solve pb solves the problem pb using GLPK with the intopt procedure. It passes arguments to the solver to guide the search.

If the function is called without parameters, the default parameters from GLPK will be used.

Generic parameters

parameter term_output
Enable/disable terminal output.

parameter msg_lev
Message level for terminal output.

Simplex parameters

parameter meth
Simplex method option.

parameter pricing
Pricing technique.

parameter r_test
Ratio test technique.

parameter it_lim
Simplex iteration limit.

parameter tm_lim
Simplex searching time limit, in milliseconds.

Mixed Integer solver parameter

parameter br_tech
Branching technique option

parameter bt_tech
Backtracking technique option

parameter pp_tech
Preprocessing technique option

parameter sr_heur
Simple rounding heuristic option

parameter fp_heur
Feasibility pump heuristic option

parameter ps_heur
Proximity search heuristic option

parameter ps_tm_lim
Time limit, in milliseconds, for the proximity search heuristic

parameter gmi_cuts
Gomory's mixed integer cut option

parameter mir_cuts
Mixed integer rounding (MIR) cut option

parameter cov_cuts
Mixed cover cut option

parameter clq_cuts
Clique cut option

parameter tol_int
Abs. tolerance to check integer feasibility

parameter tol_obj
Rel. tolerance in evaluating objective

parameter mip_gap
The relative mip gap tolerance. If the relative mip gap for currently known best integer feasible solution falls below this tolerance, the solver terminates the search. This allows obtaining suboptimal integer feasible solutions if solving the problem to optimality takes too long time.

parameter tm_lim_int
Searching time limit, in milliseconds.

returns

Ok (obj, pmap) if the problem is optimal or feasible (when search stopped because of time limit, or gap reached).

Error text if the problem is malformed, unbounded, or infeasible. The text explains the error.