Module Lp_glpk.Simplex

val solve : ?term_output:bool -> ?msg_lev:T.Msg.t option -> ?meth:T.Smcp.Meth.t option -> ?pricing:T.Smcp.Pt.t option -> ?r_test:T.Smcp.Rt.t option -> ?it_lim:int option -> ?tm_lim:int option -> Lp.Problem.t -> (float * float Lp.PMap.t, string) Stdlib.result

solve pb solves the problem pb using GLPK with the simplex algorithm. This function does not support integer or boolean variables. It passes arguments to the solver to guide the search. If the problem is optimal or feasible (when the search stopped), the found solution is returned. If the problem is malformed, unbounded, or infeasible, it returns an error.

If the function is called without parameters, the default parameters from GLPK will be used.

parameter term_output
Enable/disable terminal output.

parameter msg_lev
Message level for terminal output.

parameter meth
Simplex method option.

parameter pricing
Pricing technique.

parameter r_test
Ratio test technique.

parameter it_lim
Simplex iteration limit.

parameter tm_lim
Searching time limit, in milliseconds.

returns

Ok (obj, pmap) if the problem is optimal or feasible.

Error text if the problem is malformed, unbounded, or infeasible. The text explains the error.