Module Lp_glpk_bindings.M

Parameters

module F : Ctypes.FOREIGN

Signature

val set_term_out : (bool -> unit F.return) F.result

type prob = unit Ctypes.ptr

val prob : prob Ctypes.typ

val create_prob : (unit -> prob F.return) F.result

val delete_prob : (prob -> unit F.return) F.result

val set_prob_name : (prob -> string -> unit F.return) F.result

val get_prob_name : (prob -> string F.return) F.result

val set_obj_dir : (prob -> T.Dir.t -> unit F.return) F.result

val get_obj_dir : (prob -> T.Dir.t F.return) F.result

val add_rows : (prob -> int -> int F.return) F.result

val add_cols : (prob -> int -> int F.return) F.result

val set_row_name : (prob -> int -> string -> unit F.return) F.result

val get_row_name : (prob -> int -> string F.return) F.result

val set_col_name : (prob -> int -> string -> unit F.return) F.result

val get_col_name : (prob -> int -> string F.return) F.result

val set_row_bnds : (prob -> int -> T.Bnd.t -> float -> float -> unit F.return) F.result

set_row_bnds prob i bnd lb ub sets bounds of i-th row (constraint). If the row is not lower (upper) bounded, lb (ub) is just ignored. If the row is equality constraint (Bnd.FX), only lb is used and ub is ignored.

val set_col_bnds : (prob -> int -> T.Bnd.t -> float -> float -> unit F.return) F.result

set_col_bnds prob j bnd lb ub sets bounds of j-th col (variable). If the col is not lower (upper) bounded, lb (ub) is just ignored. If the col is equality constraint (Bnd.FX), only lb is used and ub is ignored.

val set_obj_coef : (prob -> int -> float -> unit F.return) F.result

set_obj_coef prob j sets the objective coefficient at j-th col (variable)

val set_mat_row : (prob -> int -> int -> unit Ctypes_static.ptr -> unit Ctypes_static.ptr -> unit F.return) F.result

set_mat_row prob i len indices vals sets the i-th row of constraint matrix.

val set_mat_col : (prob -> int -> int -> unit Ctypes_static.ptr -> unit Ctypes_static.ptr -> unit F.return) F.result

set_mat_col prob j len indices vals sets the j-th column of constraint matrix.

val load_matrix : (prob -> int -> unit Ctypes_static.ptr -> unit Ctypes_static.ptr -> unit Ctypes_static.ptr -> unit F.return) F.result

load_matrix prob ne ia ja ar sets the constraint matrix. The matrix is represented as an sparce matrix. for k=1 .. ne, value ark is set at (iak, jak) element.

val set_col_kind : (prob -> int -> T.Vt.t -> unit F.return) F.result

val get_col_kind : (prob -> int -> T.Vt.t F.return) F.result

val get_num_rows : (prob -> int F.return) F.result

val get_num_cols : (prob -> int F.return) F.result

val get_num_nz : (prob -> int F.return) F.result

val get_num_int : (prob -> int F.return) F.result

val get_num_bin : (prob -> int F.return) F.result

val init_smcp : (T.Smcp.t Ctypes.structure Ctypes_static.ptr -> unit F.return) F.result

val init_iocp : (T.Iocp.t Ctypes.structure Ctypes_static.ptr -> unit F.return) F.result

val simplex : (prob -> T.Smcp.t Ctypes.structure Ctypes_static.ptr -> int F.return) F.result

val intopt : (prob -> T.Iocp.t Ctypes.structure Ctypes_static.ptr -> int F.return) F.result

val get_status : (prob -> T.Stat.t F.return) F.result

val mip_status : (prob -> T.Stat.t F.return) F.result

val get_obj_val : (prob -> float F.return) F.result

val mip_obj_val : (prob -> float F.return) F.result

val get_row_prim : (prob -> int -> float F.return) F.result

val get_row_dual : (prob -> int -> float F.return) F.result

val mip_row_val : (prob -> int -> float F.return) F.result

val get_col_prim : (prob -> int -> float F.return) F.result

val get_col_dual : (prob -> int -> float F.return) F.result

val mip_col_val : (prob -> int -> float F.return) F.result